+++ gobject/tests/signals.c
@@ -1705,7 +1705,7 @@ test_clear_signal_handler (void)
 
   if (g_test_undefined ())
     {
-      handler = g_random_int_range (0x01, 0xFF);
+      handler = (gulong) g_random_int_range (0x01, 0xFF);
       g_test_expect_message (G_LOG_DOMAIN, G_LOG_LEVEL_CRITICAL,
                              "*instance '* has no handler with id *'");
       g_clear_signal_handler (&handler, test_obj);
